rename iris.cli.rpc_base -> iris.cli.connect
