+++ lib/iris/src/iris/cli/cluster.py
@@ -26,7 +26,7 @@ from iris.cli.build import (
     find_marin_root,
     get_git_sha,
 )
-from iris.cli.rpc_base import IRIS_CLUSTER_CONFIG_DIRS, require_controller_url, rpc_client
+from iris.cli.connect import IRIS_CLUSTER_CONFIG_DIRS, require_controller_url, rpc_client
 from iris.cluster.config import IrisConfig, clear_remote_state, make_local_config
 from iris.cluster.controller.autoscaler.scaling_group import (
     _zone_from_template,
